requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.21.0
